test: remove TEST_NO_CREATE_TEST_USERS env
the purpose of `TEST_NO_CREATE_TEST_USERS` was testing in local env before creating the test docker-compose.
As the test docker-compose is written, it isn't needed, addtionally, it prevents to create records in the `user` table, but not in `todo` table, so from my perspective, it can make a confusion unless the same logic is wrriten for the `todo` table, basically, all the tables.