feat: sqlmap Dockerfile
When an error occurs in database, the server was crashed. Add try-catch blocks to prevent to exit the server.

As a result, It doesn't detect any vulnerabilities for the auth and todo APIs.

If it found some vulnerabilities, I would plan to dig into it more but since it doesn't find any, finish this project. I may've done something wrong though.